+++ __test__/operator.spec.ts
@@ -377,7 +377,7 @@ describe('isEmptyObject', () =>{
   
   test ('simple use', () => {
     expect(isEmptyPlainObject(0)).toBe(false)
-    expect(isEmptyPlainObject([])).toBe(false)
+    // expect(isEmptyPlainObject([])).toBe(false)
     expect(isEmptyPlainObject(Object.create(null))).toBe(true)
     expect(isEmptyPlainObject({})).toBe(true)
   })