+++ src/dom/copyToClipboard.ts
@@ -1,22 +1,24 @@
-export function copyToClipBoard (text: string) {
+import { isWebEnv } from "../operator"
 
-  const selection = window.getSelection()
+export function copyToClipBoard (text: string) {
+  if (!isWebEnv()) return Promise.reject(1)
+  const selection = globalThis.getSelection()
   if (selection) {
-    const target = document.createElement('div')
+    const target = globalThis.document.createElement('div')
     target.style.opacity = '0'
     target.style.position = 'fixed'
     target.style.zIndex = '-1'
 
     target.innerHTML = text
-    document.body.append(target)
-    const range = document.createRange()
+    globalThis.document.body.append(target)
+    const range = globalThis.document.createRange()
 
     range.selectNode(target)
     selection.removeAllRanges()
     selection.addRange(range)
-    document.execCommand('Copy')
+    globalThis.document.execCommand('Copy')
     selection.removeAllRanges()
-    document.body.removeChild(target)
+    globalThis.document.body.removeChild(target)
 
     return Promise.resolve(0)
   }